feat: Add planning finalization and role management
Introduces the ability to finalize planning assignments for services and introduces a new `Role` type.

The `Role` type allows for more granular definition of roles within planning, differentiating between 'leader' and 'secondary' roles. This change also adds a `finalized` boolean field to the `Service` type to indicate if planning for a service has been completed.

Additionally, new icons (`LockClosedIcon`, `DownloadIcon`) and a utility function for downloading planning as a PDF have been added.